feat: Test refactors and ClientStore consolidation
There was a client store and clientrepository in the code and the two was exactly th same. I've removed ClientRepository and re-generated mocks to use the mock_client package instead of repository for handling oauth2 clients. The tests now using the client package's interface right now, but we should consider to rename the ClientStore to ClientRepository to follow the schema.